extract LogRow component
