+++ gui/src/components/logs/Logs.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { afterUpdate, beforeUpdate } from 'svelte';
 
-  import FormattedLogMessage from './FormattedLogMessage.svelte';
+  import LogRow from './LogRow.svelte';
   import type { LogEvent } from './types';
 
   export let events: LogEvent[] = [];
@@ -27,28 +27,12 @@
       logViewport.scrollTop = logViewport.scrollHeight;
     }
   });
-
-  // SEE NOTE [LOG_COMPONENTS]
-  const trimLabel = (name: string) => name.replace(/(:err$)|(:out$)/g, '');
 </script>
 
 <div class="logs-container" bind:this={logViewport}>
   <table>
     {#each events as event (event.id)}
-      <tr style={event.style}>
-        <td
-          class="time"
-          on:click={() => {
-            window.alert(`Full timestamp: ${event.time.full}`);
-          }}
-        >
-          {event.time.short}
-        </td>
-        <td class="name" title={event.name}>{trimLabel(event.name)}</td>
-        <td>
-          <FormattedLogMessage message={event.message} />
-        </td>
-      </tr>
+      <LogRow {event} />
     {/each}
   </table>
 </div>
@@ -69,42 +53,8 @@
     font-size: 15px;
   }
 
-  table,
-  tr,
-  td {
+  table {
     border: none;
     border-collapse: collapse;
-  }
-
-  td {
-    padding: 0 0.3em;
-    vertical-align: text-top;
-    color: var(--grey-3-color);
-    white-space: pre-wrap;
-  }
-
-  tr:hover td {
-    background: var(--grey-e-color);
-    color: var(--grey-1-color);
-  }
-
-  .name {
-    text-align: right;
-    background: var(--log-bg-color);
-    color: var(--log-color);
-  }
-
-  tr:hover .name {
-    background: var(--log-bg-hover-color);
-    color: var(--log-hover-color);
-  }
-
-  .time {
-    color: var(--grey-9-color);
-    cursor: zoom-in;
-  }
-
-  tr:hover .time {
-    color: var(--grey-5-color);
   }
 </style>
